requirements.txt
python-telegram-bot
pyyaml
urllib3==1.25.11 